Move DebouncingQueryTextListener to base-android module
